Alphabet Partial Implementation
Added GreekAlphabet.java file and started implementation of the two
alphabets. The user is now asked to choose between the Latin and Greek
character sets, and though the output board does not yet display the use
of the them, the correct alphabet is passed to the GameBaord file based
on the user's selection. Both LatinAlphabet and GreekAlphabet now have
char arrays of 28 chars from their respective alphabets.