Symbol Implementation
These changes give the CharacterGamePiece their correct symbols when
displaying the board. Based on the game's difficulty, the full list of
symbols is truncated, doubled in length to create pairs, and shuffled.
Then, when creating the board, each character in the list is inserted
once. LatinAlphabet was also changed so that the symbols are have been
changed from lower to upper case to better fit the specification.