Hidden symbols, user input
This commit initially hides all symbols by altering the getSymbol
method. A loop has also been created in MemoryGame that will contain the
bulk of the game - the repetition of the Turn count, the board, and the
user input prompt until the game is won. User input for selecting a tile
has been added but not yet implemented.