getRC Method Implementation
Method named getRC implemented that takes, validates, and returns a
user's input for row and column selection. A valid input is one in form
"R C" or "quit", where R and C are ints in the range of the rows and
columns on the board. If anything is invalid - the string is not in that
form or the row and column inputs are out of range - an exception is
thrown and the user is prompted again. Other methods in MemoryGame.java
have been created to simplify the MemoryGame method, nothing new has
been added to them, they've just been copy-pasted into their own
methods.