Leaderboard
This branch fully implements the leaderboard for the memory game. The
leaderboard is displayed before the game is played through the
printLeaderboard method. After the game ends, the updateLeaderboard
method is called that checks if a user broke the record. If they did,
then a congratulatory message is displayed and the user is asked their
name and the leaderboard is properly updated. If the user tied, a
similar message is displayed, but the leaderboard is not updated.
Nothing happens if the user did not break the record. The leaderboard is
stored in the leaderboard.txt file in a new resources folder.