Play Again Option
This update adds a getPlayAgain method that returns a boolean - true if
the user inputs yes, and false if no. The method also validates the
user's input and will ask again if invalid. The logic and method calls
that the game is played through and was once in MemoryGame is now in its
own playGame method. This method is called by MemoryGame and can call
itself if playAgain is true. This way, the user can play again without
being shown the welcome message each time. Comments have also been added
with a short description of each method in MemoryGame.java.